+++ spec/spec_helper.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
+require "database_cleaner/active_record"
+require "debug"
+require "factory_bot"
+
 require "propertree"
 
 RSpec.configure do |config|
@@ -12,4 +16,12 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
+
+  config.include FactoryBot::Syntax::Methods
+
+  config.before(:suite) do
+    FactoryBot.find_definitions
+  end
+
+  DatabaseCleaner.strategy = :transaction
 end